create LanguageSelector component and refactur change language to this, refactor tests
