+++ src/App.js
@@ -1,9 +1,11 @@
 import SingUpPage from "./pages/SingUpPage";
+import LanguageSelector from "./components/LanguageSelector";
 
 function App() {
   return (
       <div className="container">
         <SingUpPage/>
+        <LanguageSelector/>
       </div>
 
   );
