+++ src/pages/SingUpPage.spec.js
@@ -1,13 +1,14 @@
 import SingUpPage from './SingUpPage';
-import {render, screen, waitFor, waitForElementToBeRemoved} from '@testing-library/react';
+import {render, screen, waitFor, act, waitForElementToBeRemoved} from '@testing-library/react';
 import userEvent from "@testing-library/user-event";
 // import axios from "axios";
 import { setupServer } from "msw/node";
 import { rest } from "msw";
 // import resolve from "resolve";
-import "../locale/i18n";
+import i18n from "../locale/i18n";
 import en from "../locale/en.json";
 import pl from "../locale/pl.json";
+import LanguageSelector from "../components/LanguageSelector";
 
 let requestBody;
 let counter = 0;
@@ -213,8 +214,24 @@ describe('Sing Up Page', () => {
 
     });
     describe("Internationalization", () => {
+
+        const setup = () => {
+            render(
+                <>
+                    <SingUpPage/>
+                    <LanguageSelector/>
+                </>
+            );
+        }
+
+        afterEach(() => {
+            act(() => {
+                i18n.changeLanguage("en");
+            })
+        })
+
         it("initiality displays all text in English", () => {
-            render(<SingUpPage/>);
+            setup();
             expect(
                 screen.getByRole("heading", {name: en.signUp})
             ).toBeInTheDocument();
@@ -227,7 +244,7 @@ describe('Sing Up Page', () => {
             expect(screen.getByLabelText(en.passwordRepeat)).toBeInTheDocument();
         });
         it("displays all text in Polish after changing the Language", () => {
-            render(<SingUpPage/>);
+            setup();
 
             const polishToggle = screen.getByTitle("Polski");
             userEvent.click(polishToggle);
@@ -245,7 +262,7 @@ describe('Sing Up Page', () => {
             // nie wiem o co chodzi ale powinno być odwrotnie czyli pl nie en.
         });
         it("displays all text in English after changing back from Polish", () => {
-            render(<SingUpPage/>);
+            setup();
 
             const polishToggle = screen.getByTitle("Polski");
             userEvent.click(polishToggle);
@@ -253,16 +270,15 @@ describe('Sing Up Page', () => {
             userEvent.click(englishToggle);
 
             expect(
-                screen.getByRole("heading", {name: pl.signUp})
+                screen.getByRole("heading", {name: en.signUp})
             ).toBeInTheDocument();
             expect(
-                screen.getByRole("button", {name: pl.signUp})
+                screen.getByRole("button", {name: en.signUp})
             ).toBeInTheDocument();
-            expect(screen.getByLabelText(pl.username)).toBeInTheDocument();
-            expect(screen.getByLabelText(pl.email)).toBeInTheDocument();
-            expect(screen.getByLabelText(pl.password)).toBeInTheDocument();
-            expect(screen.getByLabelText(pl.passwordRepeat)).toBeInTheDocument();
-            // nie wiem o co chodzi ale powinno być odwrotnie czyli en nie pl.
+            expect(screen.getByLabelText(en.username)).toBeInTheDocument();
+            expect(screen.getByLabelText(en.email)).toBeInTheDocument();
+            expect(screen.getByLabelText(en.password)).toBeInTheDocument();
+            expect(screen.getByLabelText(en.passwordRepeat)).toBeInTheDocument();
         });
     });
 });
